# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RankYatra API - Competitive Exam Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── USERS ──────────────────────────────────────────────────────────────────
...
  /users/avatar:
    post:
      operationId: uploadAvatar
      tags: [users]
      summary: Upload profile picture (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadAvatarBody"
      responses:
        "200":
          description: Avatar uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarResponse"
...
  /exams:
    get:
      operationId: listExams
      tags: [exams]
      summary: List all exams (upcoming, live, completed)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, live, completed]
      responses:
        "200":
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"

    post:
      operationId: createExam
      tags: [exams]
      summary: Create a new exam (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamBody"
      responses:
        "201":
          description: Exam created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{examId}:
    get:
      operationId: getExam
      tags: [exams]
      summary: Get exam details
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateExam
      tags: [exams]
      summary: Update exam (admin only)
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExamBody"
      responses:
        "200":
          description: Updated exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"

    delete:
      operationId: deleteExam
      tags: [exams]
      summary: Delete exam (admin only)
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── QUESTIONS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ─── AUTH ─────────────────────────────────────────────────────────────────
    SignupBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    # ─── USER ─────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        walletBalance:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        isBlocked:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, walletBalance, isAdmin, isBlocked, createdAt]
...
    UploadAvatarBody:
      type: object
      properties:
        avatarBase64:
          type: string
        mimeType:
          type: string
      required: [avatarBase64, mimeType]
    AvatarResponse:
      type: object
      properties:
        avatarUrl:
          type: string
      required: [avatarUrl]
...
    Exam:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        entryFee:
          type: string
        status:
          type: string
          enum: [upcoming, live, completed]
        solutionPdfUrl:
          type: ["string", "null"]
        questionCount:
          type: integer
        registeredCount:
          type: integer
        prizePool:
          type: string
        createdAt:
          type: string
      required: [id, title, category, startTime, endTime, entryFee, status, questionCount, registeredCount, prizePool, createdAt]
    ExamDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        entryFee:
          type: string
        status:
          type: string
          enum: [upcoming, live, completed]
        solutionPdfUrl:
          type: ["string", "null"]
        questionCount:
          type: integer
        registeredCount:
          type: integer
        prizePool:
          type: string
        isRegistered:
          type: boolean
        hasSubmitted:
          type: boolean
        myResult:
          $ref: "#/components/schemas/SubmissionResult"
        createdAt:
          type: string
      required: [id, title, category, startTime, endTime, entryFee, status, questionCount, registeredCount, prizePool, isRegistered, hasSubmitted, createdAt]
    CreateExamBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        entryFee:
          type: string
        prizePool:
          type: string
        solutionPdfUrl:
          type: ["string", "null"]
      required: [title, category, startTime, endTime, entryFee, prizePool]
    UpdateExamBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        entryFee:
          type: string
        prizePool:
          type: string
        status:
          type: string
          enum: [upcoming, live, completed]
        solutionPdfUrl:
          type: ["string", "null"]

    # ─── QUESTION ────────────────────────────────────────────────────────────
...
    SubmissionResult:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        examId:
          type: integer
        score:
          type: integer
        totalQuestions:
          type: integer
        correctAnswers:
          type: integer
        timeTakenSeconds:
          type: integer
        rank:
          type: ["integer", "null"]
        submittedAt:
          type: string
      required: [id, userId, examId, score, totalQuestions, correctAnswers, timeTakenSeconds, submittedAt]

    # ─── LEADERBOARD ─────────────────────────────────────────────────────────
...